requirements.txt
# Core Framework
streamlit>=1.28.0

# AI/LLM - Groq Only
groq>=0.4.0

# Web Scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
requests>=2.31.0

# Social Media Scraping
instaloader>=4.10.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
fake-useragent>=1.4.0
